DEXW-4022: Move isSafari to common utils
